too slow tried performance upgrades
Tried solving for primes at the beginning of the run then using a text file to store and call these values when I need them. Now that I think about it I may be able to make Primes a public array and use it for all variables but It may break stuff.